cmp_2

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: ituren <ituren@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/26 17:17:24 by ituren            #+#    #+#             */
-/*   Updated: 2025/06/25 14:37:34 by ituren           ###   ########.fr       */
+/*   Updated: 2025/06/25 14:39:52 by ituren           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
